+++ cpp/src/mavsdk/core/udp_connection.cpp
@@ -231,7 +231,7 @@ void UdpConnection::add_remote_to_keep(const std::string& remote_ip, const int r
 void UdpConnection::add_remote_impl(
     const std::string& remote_ip,
     const int remote_port,
-    const uint8_t remote_sysid,
+    const uint32_t remote_sysid,
     RemoteOption remote_option)
 {
     std::lock_guard<std::mutex> lock(_remote_mutex);
@@ -290,7 +290,7 @@ void UdpConnection::do_receive()
             auto parse_result = _mavlink_receiver->parse_message();
             while (parse_result != MavlinkReceiver::ParseResult::NoneAvailable) {
                 if (parse_result == MavlinkReceiver::ParseResult::MessageParsed) {
-                    const uint8_t sysid = _mavlink_receiver->get_last_message().sysid;
+                    const uint32_t sysid = _mavlink_receiver->get_last_message().sysid;
                     if (sysid != 0) {
                         add_remote_impl(
                             _sender_endpoint.address().to_string(),
